requirements.txt
# MT5/MT4 Trading HTTP Server Dependencies

# Core web framework
Flask
Flask-CORS

# MT5/MT4 connection
MetaTrader5
MetaTrader4

# Data handling
pandas
numpy

# HTTP requests (for testing)
requests

# Configuration and logging
python-dotenv
PyYAML

# Date and time handling
pytz

# JSON schema validation
jsonschema

# Production WSGI server
waitress
